feat: ticket templates system — tabbed builder, attachments, user pickers, V31-V34 migrations
- V31: attachments table (content-addressed file storage, token-gated access)
- V32: ticket_templates + ticket_template_versions tables (tabbed JSON layout)
- V33: TICKET_MANAGER permission for responsible field picker
- V34: migrate old flat-array layouts to tabbed format (fixes deserialization crash)
- TemplateEntity, TemplateVersionEntity, TemplateRepository, TemplateVersionRepository
- TemplateService: tabbed layout, versioning, AI suggest, buildDefaultLayout()
- TemplateController: CRUD + AI suggest endpoint
- TicketUserController: /assignable + /managers (TICKET_MANAGER holders)
- UserSelectItemDto: user_id + display_name
- Repository additions: findByPermissionKey, findByPermissionId, findByGroupIdIn
- AttachmentController/Service/Entity/Dto: updated for token-gated preview/stream/download

+++ src/main/java/com/turbotikects/turbotikectsserver/dto/AttachmentDto.java
@@ -12,6 +12,6 @@ public class AttachmentDto {
     private String originalFilename;
     private String mimeType;
     private Long fileSize;
-    private Long uploadedBy;
+    private Integer uploadedBy;
     private LocalDateTime createdAt;
 }
